al buscar una partida por matchId, primero busca en la base de datos. tuve que modificar algun que otro dto.
diosito, perdon por como hice la conversion de las fechas

+++ src/main/java/riot/riotapi/dtos/match/ParticipantDTO.java
@@ -18,7 +18,7 @@ public class ParticipantDTO {
     private ChallengesDTO challenges;
     private Integer champExperience;
     private Integer champLevel;
-    private Integer championId;
+    private Long championId;
     private String championName;
     private Integer championTransform;
     private Integer commandPings;
